Innotter/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
attrs==22.1.0; python_version >= '3.5'
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.0; python_version >= '3.6'
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.13.0
django==3.2.15
djangorestframework==3.13.1
drf-yasg==1.21.3
factory-boy==3.2.1; python_version >= '3.6'
faker==14.0.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
model-bakery==1.7.0
packaging==21.3; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
psycopg2==2.9.3
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyjwt==2.4.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest==7.1.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2.1
requests==2.28.1; python_version >= '3.7' and python_version < '4'
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tomli==2.0.1; python_version >= '3.7'
typing-extensions==4.3.0; python_version >= '3.7'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
